refactor: web 관련 클래스 common -> api 이동
