+++ src/main/kotlin/dnd11th/blooming/api/annotation/ApiErrorResponse.kt
@@ -1,4 +1,4 @@
-package dnd11th.blooming.common.annotation
+package dnd11th.blooming.api.annotation
 
 import dnd11th.blooming.common.exception.ErrorType
