+++ src/main/kotlin/dnd11th/blooming/api/annotation/PendingUser.kt
@@ -1,4 +1,4 @@
-package dnd11th.blooming.common.annotation
+package dnd11th.blooming.api.annotation
 
 @Retention(AnnotationRetention.RUNTIME)
 @Target(AnnotationTarget.VALUE_PARAMETER)
