+++ src/main/kotlin/dnd11th/blooming/api/controller/home/HomeController.kt
@@ -2,8 +2,8 @@ package dnd11th.blooming.api.controller.home
 
 import dnd11th.blooming.api.dto.home.HomeResponse
 import dnd11th.blooming.api.service.home.HomeService
-import dnd11th.blooming.common.annotation.LoginUser
-import dnd11th.blooming.common.annotation.Secured
+import dnd11th.blooming.api.annotation.LoginUser
+import dnd11th.blooming.api.annotation.Secured
 import dnd11th.blooming.domain.entity.user.User
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
