+++ src/main/kotlin/dnd11th/blooming/api/controller/image/ImageController.kt
@@ -3,8 +3,8 @@ package dnd11th.blooming.api.controller.image
 import dnd11th.blooming.api.dto.image.ImageFavoriteModifyRequest
 import dnd11th.blooming.api.dto.image.ImageSaveRequest
 import dnd11th.blooming.api.service.image.ImageService
-import dnd11th.blooming.common.annotation.LoginUser
-import dnd11th.blooming.common.annotation.Secured
+import dnd11th.blooming.api.annotation.LoginUser
+import dnd11th.blooming.api.annotation.Secured
 import dnd11th.blooming.domain.entity.user.User
 import jakarta.validation.Valid
 import org.springframework.web.bind.annotation.DeleteMapping
