+++ src/main/kotlin/dnd11th/blooming/api/controller/weather/WeatherMessageApi.kt
@@ -1,7 +1,7 @@
 package dnd11th.blooming.api.controller.weather
 
 import dnd11th.blooming.api.dto.weather.WeatherMessageResponse
-import dnd11th.blooming.common.annotation.ApiErrorResponse
+import dnd11th.blooming.api.annotation.ApiErrorResponse
 import dnd11th.blooming.common.exception.ErrorType
 import dnd11th.blooming.domain.entity.user.User
 import io.swagger.v3.oas.annotations.Operation
