+++ src/main/kotlin/dnd11th/blooming/api/controller/weather/WeatherMessageController.kt
@@ -3,8 +3,8 @@ package dnd11th.blooming.api.controller.weather
 import dnd11th.blooming.api.dto.weather.WeatherMessageResponse
 import dnd11th.blooming.api.service.weather.WeatherMessage
 import dnd11th.blooming.api.service.weather.WeatherService
-import dnd11th.blooming.common.annotation.LoginUser
-import dnd11th.blooming.common.annotation.Secured
+import dnd11th.blooming.api.annotation.LoginUser
+import dnd11th.blooming.api.annotation.Secured
 import dnd11th.blooming.domain.entity.user.User
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
