+++ src/main/kotlin/dnd11th/blooming/api/resolver/PendingUserArgumentResolver.kt
@@ -1,7 +1,7 @@
-package dnd11th.blooming.common.resolver
+package dnd11th.blooming.api.resolver
 
-import dnd11th.blooming.common.annotation.PendingUser
-import dnd11th.blooming.common.jwt.JwtProvider
+import dnd11th.blooming.api.annotation.PendingUser
+import dnd11th.blooming.api.jwt.JwtProvider
 import dnd11th.blooming.domain.entity.user.RegisterClaims
 import jakarta.servlet.http.HttpServletRequest
 import org.springframework.core.MethodParameter
