+++ src/main/kotlin/dnd11th/blooming/api/service/user/LogoutService.kt
@@ -1,6 +1,6 @@
 package dnd11th.blooming.api.service.user
 
-import dnd11th.blooming.common.jwt.JwtProvider
+import dnd11th.blooming.api.jwt.JwtProvider
 import dnd11th.blooming.domain.entity.user.UserClaims
 import dnd11th.blooming.domain.repository.token.RefreshTokenRepository
 import org.springframework.stereotype.Service
