+++ src/main/kotlin/dnd11th/blooming/api/service/user/UserRegisterService.kt
@@ -3,7 +3,7 @@ package dnd11th.blooming.api.service.user
 import dnd11th.blooming.api.dto.user.TokenResponse
 import dnd11th.blooming.common.exception.ErrorType
 import dnd11th.blooming.common.exception.NotFoundException
-import dnd11th.blooming.common.jwt.JwtProvider
+import dnd11th.blooming.api.jwt.JwtProvider
 import dnd11th.blooming.domain.entity.refreshtoken.RefreshToken
 import dnd11th.blooming.domain.entity.region.Region
 import dnd11th.blooming.domain.entity.user.OauthProvider
