+++ src/test/kotlin/dnd11th/blooming/common/jwt/JwtProviderTest.kt
@@ -1,5 +1,7 @@
 package dnd11th.blooming.common.jwt
 
+import dnd11th.blooming.api.jwt.JwtProperties
+import dnd11th.blooming.api.jwt.JwtProvider
 import dnd11th.blooming.domain.entity.user.UserClaims
 import io.jsonwebtoken.Claims
 import io.jsonwebtoken.Jwts
